+++ ex02/Account.hpp
@@ -11,6 +11,9 @@
 #pragma once
 #ifndef __ACCOUNT_H__
 #define __ACCOUNT_H__
+#include <iostream>
+#include <time.h>
+#include <iomanip>
 
 // ************************************************************************** //
 //                               Account Class                                //
